requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
rembg==2.0.57
Pillow==10.2.0
numpy==1.26.4
opencv-python==4.9.0.80
gunicorn==21.2.0
python-dotenv==1.0.1